Move custom scaleOptions to the image block
